Improve structure and documentation of Hierarchical connector
